Resolve focused S08 review findings
