+++ .supportability-characterization.json
@@ -28,9 +28,17 @@
     },
     {
       "covers": [
+        "src/supportability_gate/standard_results.py",
+        "src/supportability_gate/standard_results_enforcer.py",
+        "src/supportability_gate/standard_results_producer.py"
+      ],
+      "id": "standard-results-boundary",
+      "kind": "regression"
+    },
+    {
+      "covers": [
         "src/supportability_gate/quality_profile.py",
-        "src/supportability_gate/reporting.py",
-        "src/supportability_gate/standard_results.py"
+        "src/supportability_gate/reporting.py"
       ],
       "id": "m9-quality-profile",
       "kind": "cli"
