+++ src/supportability_gate/standard_results.py
@@ -1,15 +1,11 @@
-"""Compose and enforce one independently visible result per Supportability Standard."""
+"""Compose one independently visible result per Supportability Standard."""
 
 from __future__ import annotations
 
-import argparse
 import hashlib
 import json
-import os
 import re
 from dataclasses import dataclass
-from datetime import datetime
-from pathlib import Path
 from typing import Any
 
 from supportability_gate import clause_inventory, codex_review
@@ -208,16 +204,6 @@ class DeterministicState:
 
 def _canonical(value: object) -> bytes:
     return json.dumps(value, ensure_ascii=False, separators=(",", ":"), sort_keys=True).encode()
-
-
-def _read_json(path: Path, code: str) -> dict[str, Any]:
-    try:
-        value = json.loads(path.read_bytes())
-    except (OSError, json.JSONDecodeError) as error:
-        raise StandardResultsError(code) from error
-    if not isinstance(value, dict):
-        raise StandardResultsError(code)
-    return value
 
 
 def _string_list(value: object, code: str) -> list[str]:
@@ -610,311 +596,4 @@ def compose_results(
         codex_blocks,
         codex_technical,
     )
-    validate_payload(payload, identity)
     return payload
-
-
-def _time(value: object) -> datetime:
-    if not isinstance(value, str):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    try:
-        parsed = datetime.fromisoformat(value.replace("Z", "+00:00"))
-    except ValueError as error:
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY") from error
-    if parsed.tzinfo is None:
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    return parsed
-
-
-def _completion_identity(
-    completion: object, request_id: int | None, requested_time: datetime | None
-) -> tuple[tuple[str, int] | None, datetime | None]:
-    if completion is None:
-        return None, None
-    if (
-        not isinstance(completion, dict)
-        or set(completion) != {"completed_at", "id", "kind"}
-        or type(completion.get("id")) is not int
-        or completion["id"] < 1
-        or completion.get("kind") not in {"comment", "reaction", "review"}
-        or not isinstance(completion.get("completed_at"), str)
-    ):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    if request_id is None or requested_time is None:
-        raise StandardResultsError("MISSING_STANDARD_CODEX_BINDING")
-    completed_time = _time(completion["completed_at"])
-    if completed_time < requested_time:
-        raise StandardResultsError("STANDARD_CODEX_BINDING_MISMATCH")
-    return (completion["kind"], completion["id"]), completed_time
-
-
-def _validate_codex(
-    value: object, focus: str, result: str, blocks: list[str]
-) -> tuple[int | None, datetime | None, tuple[str, int] | None, datetime | None]:
-    expected = {"completion", "focus", "request_id", "requested_at"}
-    if not isinstance(value, dict) or set(value) != expected or value.get("focus") != focus:
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    request_id, requested_at = value.get("request_id"), value.get("requested_at")
-    if request_id is not None and (type(request_id) is not int or request_id < 1):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    if (request_id is None) != (requested_at is None):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    requested_time = _time(requested_at) if requested_at is not None else None
-    artifact, completed_time = _completion_identity(
-        value.get("completion"), request_id, requested_time
-    )
-    codex_blocks = {
-        f"FOCUSED_CODEX_REVIEW_PENDING_{focus}",
-        f"MISSING_FOCUSED_CODEX_REVIEW_REQUEST_{focus}",
-    } & set(blocks)
-    if result == "PASS" and artifact is None:
-        raise StandardResultsError("MISSING_STANDARD_CODEX_BINDING")
-    if result == "BLOCK" and artifact is None and not codex_blocks:
-        raise StandardResultsError("MISSING_STANDARD_CODEX_BINDING")
-    if artifact is not None and codex_blocks:
-        raise StandardResultsError("STANDARD_CODEX_BINDING_MISMATCH")
-    return request_id, requested_time, artifact, completed_time
-
-
-def _validate_entry(
-    value: object, standard: int
-) -> tuple[int | None, datetime | None, tuple[str, int] | None, datetime | None]:
-    expected = {
-        "applicable",
-        "blocks",
-        "check_context",
-        "codex_review",
-        "evidence_sources",
-        "result",
-        "standard",
-        "technical_errors",
-    }
-    if not isinstance(value, dict) or set(value) != expected:
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    result = value.get("result")
-    if (
-        value.get("standard") != standard
-        or value.get("check_context") != CHECK_CONTEXTS[standard - 1]
-        or type(value.get("applicable")) is not bool
-        or result not in RESULTS
-        or value.get("evidence_sources") != list(EVIDENCE_SOURCES[standard - 1])
-    ):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    blocks = _string_list(value.get("blocks"), "MALFORMED_STANDARD_RESULT_ENTRY")
-    technical = _string_list(value.get("technical_errors"), "MALFORMED_STANDARD_RESULT_ENTRY")
-    if blocks != sorted(blocks) or technical != sorted(technical):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    codex_families = {
-        f"FOCUSED_CODEX_REVIEW_PENDING_{standard}",
-        f"MISSING_FOCUSED_CODEX_REVIEW_REQUEST_{standard}",
-    }
-    for block in set(blocks) - codex_families:
-        _require_owner(block, standard)
-    if result == "PASS" and (blocks or technical):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    if result == "BLOCK" and (not blocks or technical):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    if result == "TECHNICAL_FAILURE" and not technical:
-        raise StandardResultsError("MALFORMED_STANDARD_RESULT_ENTRY")
-    return _validate_codex(value.get("codex_review"), str(standard), str(result), blocks)
-
-
-def _validate_quality_artifact(value: object, technical: bool) -> None:
-    if value is None and technical:
-        return
-    if (
-        not isinstance(value, dict)
-        or set(value) != {"capture_sha256", "digest", "id"}
-        or type(value.get("id")) is not int
-        or value["id"] < 1
-        or not isinstance(value.get("digest"), str)
-        or SHA64.fullmatch(value["digest"]) is None
-        or not isinstance(value.get("capture_sha256"), str)
-        or SHA64.fullmatch(value["capture_sha256"]) is None
-    ):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULTS_ARTIFACT")
-
-
-def validate_payload(value: object, identity: RunIdentity | None = None) -> dict[str, Any]:
-    """Validate ordering, identities, bindings, and every lane result."""
-    expected = {
-        "base_sha",
-        "entries",
-        "head_sha",
-        "quality_artifact",
-        "repository",
-        "repository_id",
-        "run_attempt",
-        "run_id",
-        "schema_version",
-        "standard_sha256",
-        "workflow_sha",
-    }
-    if not isinstance(value, dict) or set(value) != expected:
-        raise StandardResultsError("MALFORMED_STANDARD_RESULTS")
-    repository_id = value.get("repository_id")
-    run_id = value.get("run_id")
-    run_attempt = value.get("run_attempt")
-    actual = RunIdentity(
-        str(value.get("repository")),
-        repository_id if type(repository_id) is int else 0,
-        str(value.get("base_sha")),
-        str(value.get("head_sha")),
-        str(value.get("workflow_sha")),
-        run_id if type(run_id) is int else 0,
-        run_attempt if type(run_attempt) is int else 0,
-    )
-    _validate_identity(actual)
-    if identity is not None and actual != identity:
-        raise StandardResultsError("STANDARD_RESULTS_BINDING_MISMATCH")
-    entries = value.get("entries")
-    if (
-        value.get("schema_version") != SCHEMA_VERSION
-        or value.get("standard_sha256") != clause_inventory.STANDARD_SHA256
-        or not isinstance(entries, list)
-        or len(entries) != 8
-    ):
-        raise StandardResultsError("MALFORMED_STANDARD_RESULTS")
-    codex = [_validate_entry(entry, standard) for standard, entry in enumerate(entries, start=1)]
-    request_ids = [item[0] for item in codex if item[0] is not None]
-    artifacts = [item[2] for item in codex if item[2] is not None]
-    if len(request_ids) != len(set(request_ids)) or len(artifacts) != len(set(artifacts)):
-        raise StandardResultsError("REUSED_FOCUSED_CODEX_REVIEW_EVIDENCE")
-    for current, following in zip(codex, codex[1:], strict=False):
-        if (current[1] is not None and following[1] is not None and current[1] >= following[1]) or (
-            current[3] is not None and following[1] is not None and current[3] >= following[1]
-        ):
-            raise StandardResultsError("OUT_OF_ORDER_FOCUSED_CODEX_REVIEW_EVIDENCE")
-    technical_rows = [entry.get("result") == "TECHNICAL_FAILURE" for entry in entries]
-    if any(technical_rows) != all(technical_rows):
-        raise StandardResultsError("INCONSISTENT_SHARED_TECHNICAL_FAILURE")
-    technical = all(technical_rows)
-    if technical and len({tuple(entry["technical_errors"]) for entry in entries}) != 1:
-        raise StandardResultsError("INCONSISTENT_SHARED_TECHNICAL_FAILURE")
-    _validate_quality_artifact(value.get("quality_artifact"), technical)
-    return value
-
-
-def _write_json(path: Path, value: object) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_bytes(
-        json.dumps(value, ensure_ascii=False, indent=2, sort_keys=True).encode() + b"\n"
-    )
-
-
-def _identity(arguments: argparse.Namespace) -> RunIdentity:
-    return RunIdentity(
-        arguments.repository,
-        arguments.repository_id,
-        arguments.base_sha,
-        arguments.head_sha,
-        arguments.workflow_sha,
-        arguments.run_id,
-        arguments.run_attempt,
-    )
-
-
-def _compose(arguments: argparse.Namespace) -> int:
-    identity = _identity(arguments)
-    try:
-        complexity = _read_json(Path(arguments.complexity_result), "MISSING_COMPLEXITY_RESULT")
-        characterization = _read_json(
-            Path(arguments.characterization_result), "MISSING_CHARACTERIZATION_RESULT"
-        )
-        refactor = _read_json(Path(arguments.refactor_result), "MISSING_REFACTOR_RESULT")
-        quality = _read_json(Path(arguments.quality_provenance), "MISSING_QUALITY_PROVENANCE")
-    except StandardResultsError as error:
-        payload = compose_results(
-            {}, {}, {}, {}, identity, (), codex_review.CodexReviewError(error.code)
-        )
-    else:
-        token = os.environ.get("GITHUB_TOKEN")
-        evidence: tuple[codex_review.FocusedReviewEvidence, ...] = ()
-        codex_error: codex_review.CodexReviewError | None = None
-        if not token:
-            codex_error = codex_review.CodexReviewError("GITHUB_CODEX_REVIEW_EVIDENCE_FAILURE")
-        else:
-            try:
-                evidence = codex_review.require_focused_completion(
-                    identity.repository,
-                    arguments.pull_number,
-                    identity.head_sha,
-                    identity.run_id,
-                    token,
-                )
-            except codex_review.CodexReviewError as error:
-                codex_error = error
-                try:
-                    _, evidence = codex_review.focused_completion_snapshot(
-                        identity.repository,
-                        arguments.pull_number,
-                        identity.head_sha,
-                        identity.run_id,
-                        token,
-                    )
-                except codex_review.CodexReviewError as snapshot_error:
-                    evidence = ()
-                    codex_error = snapshot_error
-        payload = compose_results(
-            complexity,
-            characterization,
-            refactor,
-            quality,
-            identity,
-            evidence,
-            codex_error,
-        )
-    _write_json(Path(arguments.output), payload)
-    print("STANDARD_RESULTS_EMITTED")
-    return 0
-
-
-def _enforce(arguments: argparse.Namespace) -> int:
-    try:
-        payload = validate_payload(
-            _read_json(Path(arguments.input), "MISSING_STANDARD_RESULTS"), _identity(arguments)
-        )
-        entry = payload["entries"][arguments.standard - 1]
-    except (StandardResultsError, IndexError) as error:
-        print(getattr(error, "code", "MALFORMED_STANDARD_RESULTS"))
-        return 2
-    print(json.dumps(entry, ensure_ascii=False, sort_keys=True))
-    return {"PASS": 0, "BLOCK": 1, "TECHNICAL_FAILURE": 2}[entry["result"]]
-
-
-def _identity_arguments(parser: argparse.ArgumentParser) -> None:
-    parser.add_argument("--repository", required=True)
-    parser.add_argument("--repository-id", required=True, type=int)
-    parser.add_argument("--base-sha", required=True)
-    parser.add_argument("--head-sha", required=True)
-    parser.add_argument("--workflow-sha", required=True)
-    parser.add_argument("--run-id", required=True, type=int)
-    parser.add_argument("--run-attempt", required=True, type=int)
-
-
-def _parser() -> argparse.ArgumentParser:
-    parser = argparse.ArgumentParser(prog="supportability-standard-results")
-    commands = parser.add_subparsers(dest="command", required=True)
-    compose = commands.add_parser("compose")
-    _identity_arguments(compose)
-    compose.add_argument("--pull-number", required=True, type=int)
-    compose.add_argument("--complexity-result", required=True)
-    compose.add_argument("--characterization-result", required=True)
-    compose.add_argument("--refactor-result", required=True)
-    compose.add_argument("--quality-provenance", required=True)
-    compose.add_argument("--output", required=True)
-    enforce = commands.add_parser("enforce")
-    _identity_arguments(enforce)
-    enforce.add_argument("--input", required=True)
-    enforce.add_argument("--standard", required=True, type=int, choices=range(1, 9))
-    return parser
-
-
-def main(argv: list[str] | None = None) -> int:
-    """Compose or enforce standard results at an isolated workflow boundary."""
-    arguments = _parser().parse_args(argv)
-    return _compose(arguments) if arguments.command == "compose" else _enforce(arguments)
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())
